+++ modules/kubernetes/00-variables_defaults.tf
@@ -1,7 +1,7 @@
 variable "ami_name_filter_regex" {
   type        = string
   description = "Regex to find the ami to use"
-  default     = "ubuntu/images/hvm-ssd/ubuntu-bionic-18.04-amd64-server-*"
+  default     = "ubuntu/images/hvm-ssd/ubuntu-focal-20.04-amd64-server-*"
 }
 
 variable "ami_owner" {
@@ -74,4 +74,16 @@ variable "sns_topic_notifications" {
   type        = string
   description = "The SNS topic to notify when the system autoscale. If empty no notification will be sent"
   default     = ""
+}
+
+variable "k8s_deb_package_version" {
+  type        = string
+  description = "The version of the deb package to install in ubuntu (i.e. 1.18.2)"
+  default     = "1.18.2"
+}
+
+variable "kubeadm_install_version" {
+  type        = string
+  description = "The version to install in the syntax expected by kubeadm (i.e. stable-1.18)"
+  default     = "stable-1.18"
 }